+++ CHANGELOG.md
@@ -8,6 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add `nolazyloading` attribute to turn off lazy loading on particular image or video.
+
+### Changed
+
+- Upgrade dependencies.
+
 ## [0.11.0] - 2022-09-29
 
 ### Added