+++ CHANGELOG.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 
 - Eliminates excess slashes in the path and domain values configured when setting up components
-- Upgrade dependencies (`svelte.js` up to 3.52.0)
+- Upgrade dependencies (`Next.js` up to 13.0.0)
+- Upgrade dependencies (`Svelte.js` up to 3.52.0)
 
 ## [0.12.0] - 2022-10-12