+++ CHANGELOG.md
@@ -8,9 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.13.0] - 2022-11-10
+
 ### Added
 
-- Add `TwicImg` support for **react-native**
+- Add `TwicImg` support for **React Native**
 
 ### Changed
 
@@ -305,6 +307,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support for **React**, **Next**, **Vue 2**, **Vue 3**, and **Nuxt 2**
 
 [Unreleased]: https://github.com/TwicPics/components/compare/main...dev
+[0.13.0]: https://github.com/TwicPics/components/compare/0.12.0...0.13.0
 [0.12.0]: https://github.com/TwicPics/components/compare/0.11.0...0.12.0
 [0.11.0]: https://github.com/TwicPics/components/compare/0.10.0...0.11.0
 [0.10.0]: https://github.com/TwicPics/components/compare/0.9.1...0.10.0